requirements.txt
Flask
Flask-SQLAlchemy
flask-login
flask-bcrypt
psycopg2-binary
